-changed realtime property in EnergyDto to double -create mapper for energymetertosimpledto -added energymeterservice and added service to controller

+++ TecEnergy.Database/Models/DtoModels/EnergyDto.cs
@@ -9,6 +9,6 @@ public class EnergyDto
 {
     public Guid Id { get; set; }
     public string Name { get; set; }
-    public int RealTime { get; set; }
+    public double RealTime { get; set; }
     public long Accumulated { get; set; }
 }
